Split out client modules
