+++ src/client/mod.rs
@@ -1,5 +1,10 @@
 #![allow(warnings, missing_debug_implementations)]
 
+pub mod unary;
+pub mod client_streaming;
+pub mod server_streaming;
+pub mod streaming;
+
 use Status;
 
 use futures::{stream, Future, Stream, Poll};
@@ -208,218 +213,6 @@ impl Builder {
             inner,
             scheme,
             authority,
-        }
-    }
-}
-
-pub mod unary {
-    use super::client_streaming;
-    use codec::Streaming;
-
-    use futures::{Future, Stream, Poll};
-    use http::{response, Response};
-    use prost::Message;
-    use tower_h2::{Body, Data};
-
-    use std::marker::PhantomData;
-
-    pub struct ResponseFuture<T, U, B> {
-        inner: client_streaming::ResponseFuture<T, U, B>,
-    }
-
-    impl<T, U, B> ResponseFuture<T, U, B> {
-        /// Create a new client-streaming response future.
-        pub(crate) fn new(inner: client_streaming::ResponseFuture<T, U, B>) -> Self {
-            ResponseFuture { inner }
-        }
-    }
-
-    impl<T, U, B> Future for ResponseFuture<T, U, B>
-    where T: Message + Default,
-          U: Future<Item = Response<B>>,
-          B: Body<Data = Data>,
-    {
-        type Item = ::Response<T>;
-        type Error = ::Error<U::Error>;
-
-        fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-            self.inner.poll()
-        }
-    }
-}
-
-pub mod client_streaming {
-    use super::streaming;
-    use codec::Streaming;
-
-    use futures::{Future, Stream, Poll};
-    use http::{response, Response};
-    use prost::Message;
-    use tower_h2::{Body, Data};
-
-    use std::marker::PhantomData;
-
-    pub struct ResponseFuture<T, U, B> {
-        state: State<T, U, B>,
-    }
-
-    enum State<T, U, B> {
-        WaitResponse(streaming::ResponseFuture<T, U>),
-        WaitMessage {
-            head: Option<response::Parts>,
-            stream: Streaming<T, B>,
-        },
-    }
-
-    impl<T, U, B> ResponseFuture<T, U, B> {
-        /// Create a new client-streaming response future.
-        pub(crate) fn new(inner: streaming::ResponseFuture<T, U>) -> Self {
-            let state = State::WaitResponse(inner);
-            ResponseFuture { state }
-        }
-    }
-
-    impl<T, U, B> Future for ResponseFuture<T, U, B>
-    where T: Message + Default,
-          U: Future<Item = Response<B>>,
-          B: Body<Data = Data>,
-    {
-        type Item = ::Response<T>;
-        type Error = ::Error<U::Error>;
-
-        fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-            use self::State::*;
-
-            loop {
-                let response = match self.state {
-                    WaitResponse(ref mut inner) => {
-                        try_ready!(inner.poll())
-                    }
-                    WaitMessage { ref mut head, ref mut stream } => {
-                        let res = stream.poll()
-                            .map_err(|e| {
-                                match e {
-                                    ::Error::Grpc(s) => ::Error::Grpc(s),
-                                    _ => ::Error::Grpc(::Status::INTERNAL),
-                                }
-                            });
-
-                        let message = match try_ready!(res) {
-                            Some(message) => message,
-                            // TODO: handle missing message
-                            None => unimplemented!(),
-                        };
-
-                        let head = head.take().unwrap();
-                        let response = Response::from_parts(head, message);
-
-                        return Ok(::Response::from_http(response).into());
-                    }
-                };
-
-                let (head, body) = response
-                    .into_http()
-                    .into_parts();
-
-                self.state = WaitMessage {
-                    head: Some(head),
-                    stream: body,
-                };
-            }
-        }
-    }
-}
-
-pub mod server_streaming {
-    use super::streaming;
-    use codec::Streaming;
-
-    use futures::{Future, Stream, Poll};
-    use http::{response, Response};
-    use prost::Message;
-    use tower_h2::{Body, Data};
-
-    use std::marker::PhantomData;
-
-    pub struct ResponseFuture<T, U> {
-        inner: streaming::ResponseFuture<T, U>,
-    }
-
-    impl<T, U> ResponseFuture<T, U> {
-        /// Create a new client-streaming response future.
-        pub(crate) fn new(inner: streaming::ResponseFuture<T, U>) -> Self {
-            ResponseFuture { inner }
-        }
-    }
-
-    impl<T, U, B> Future for ResponseFuture<T, U>
-    where T: Message + Default,
-          U: Future<Item = Response<B>>,
-          B: Body<Data = Data>,
-    {
-        type Item = ::Response<Streaming<T, B>>;
-        type Error = ::Error<U::Error>;
-
-        fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-            self.inner.poll()
-        }
-    }
-}
-
-pub mod streaming {
-    use codec::Streaming;
-
-    use futures::{Future, Poll};
-    use http::Response;
-    use prost::Message;
-    use tower_h2::{Body, Data};
-
-    use std::marker::PhantomData;
-
-    pub struct ResponseFuture<T, U> {
-        inner: U,
-        _m: PhantomData<T>,
-    }
-
-    impl<T, U> ResponseFuture<T, U> {
-        /// Create a new client-streaming response future.
-        pub(crate) fn new(inner: U) -> Self {
-            ResponseFuture {
-                inner,
-                _m: PhantomData,
-            }
-        }
-    }
-
-    impl<T, U, B> Future for ResponseFuture<T, U>
-    where T: Message + Default,
-          U: Future<Item = Response<B>>,
-          B: Body<Data = Data>,
-    {
-        type Item = ::Response<Streaming<T, B>>;
-        type Error = ::Error<U::Error>;
-
-        fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-            use codec::Decoder;
-            use generic::Streaming;
-
-            let response = self.inner.poll()
-                .map_err(::Error::Inner);
-
-            // Get the response
-            let response = try_ready!(response);
-
-            // Destructure into the head / body
-            let (head, body) = response.into_parts();
-
-            if let Some(status) = super::check_grpc_status(&head.headers) {
-                return Err(::Error::Grpc(status));
-            }
-
-            let body = Streaming::new(Decoder::new(), body, true);
-            let response = Response::from_parts(head, body);
-
-            Ok(::Response::from_http(response).into())
         }
     }
 }
