+++ src/client/unary.rs
@@ -1,0 +1,33 @@
+use super::client_streaming;
+use codec::Streaming;
+
+use futures::{Future, Stream, Poll};
+use http::{response, Response};
+use prost::Message;
+use tower_h2::{Body, Data};
+
+use std::marker::PhantomData;
+
+pub struct ResponseFuture<T, U, B> {
+    inner: client_streaming::ResponseFuture<T, U, B>,
+}
+
+impl<T, U, B> ResponseFuture<T, U, B> {
+    /// Create a new client-streaming response future.
+    pub(crate) fn new(inner: client_streaming::ResponseFuture<T, U, B>) -> Self {
+        ResponseFuture { inner }
+    }
+}
+
+impl<T, U, B> Future for ResponseFuture<T, U, B>
+where T: Message + Default,
+      U: Future<Item = Response<B>>,
+      B: Body<Data = Data>,
+{
+    type Item = ::Response<T>;
+    type Error = ::Error<U::Error>;
+
+    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
+        self.inner.poll()
+    }
+}
